+++ src/pd/mod.rs
@@ -13,12 +13,11 @@
 
 use std::vec::Vec;
 
+use futures;
 use futures::Future;
 use futures::BoxFuture;
 
 mod async;
-// TODO: use metrics.
-#[allow(dead_code)]
 mod metrics;
 
 pub mod errors;
@@ -88,9 +87,7 @@ pub trait PdClient: Send + Sync {
     fn get_region(&self, key: &[u8]) -> Result<metapb::Region>;
 
     // Get region by region id.
-    fn get_region_by_id(&self, region_id: u64) -> Result<Option<metapb::Region>> {
-        self.get_region_by_id_async(region_id).wait()
-    }
+    fn get_region_by_id(&self, region_id: u64) -> Result<Option<metapb::Region>>;
 
     // Leader for a region will use this to heartbeat Pd.
     fn region_heartbeat(&self,
@@ -99,43 +96,20 @@ pub trait PdClient: Send + Sync {
                         down_peers: Vec<pdpb::PeerStats>,
                         pending_peers: Vec<metapb::Peer>,
                         written_bytes: u64)
-                        -> Result<pdpb::RegionHeartbeatResponse> {
-        self.region_heartbeat_async(region, leader, down_peers, pending_peers, written_bytes).wait()
-    }
+                        -> Result<pdpb::RegionHeartbeatResponse>;
 
     // Ask pd for split, pd will returns the new split region id.
-    fn ask_split(&self, region: metapb::Region) -> Result<pdpb::AskSplitResponse> {
-        self.ask_split_async(region).wait()
-    }
+    fn ask_split(&self, region: metapb::Region) -> Result<pdpb::AskSplitResponse>;
 
     // Send store statistics regularly.
-    fn store_heartbeat(&self, stats: pdpb::StoreStats) -> Result<()> {
-        self.store_heartbeat_async(stats).wait()
-    }
+    fn store_heartbeat(&self, stats: pdpb::StoreStats) -> Result<()>;
 
     // Report pd the split region.
-    fn report_split(&self, left: metapb::Region, right: metapb::Region) -> Result<()> {
-        self.report_split_async(left, right).wait()
-    }
+    fn report_split(&self, left: metapb::Region, right: metapb::Region) -> Result<()>;
 
     // Get region by region id.
-    fn get_region_by_id_async(&self, region_id: u64) -> PdFuture<Option<metapb::Region>>;
-
-    // Leader for a region will use this to heartbeat Pd.
-    fn region_heartbeat_async(&self,
-                              region: metapb::Region,
-                              leader: metapb::Peer,
-                              down_peers: Vec<pdpb::PeerStats>,
-                              pending_peers: Vec<metapb::Peer>,
-                              written_bytes: u64)
-                              -> PdFuture<pdpb::RegionHeartbeatResponse>;
-
-    // Ask pd for split, pd will returns the new split region id.
-    fn ask_split_async(&self, region: metapb::Region) -> PdFuture<pdpb::AskSplitResponse>;
-
-    // Send store statistics regularly.
-    fn store_heartbeat_async(&self, stats: pdpb::StoreStats) -> PdFuture<()>;
-
-    // Report pd the split region.
-    fn report_split_async(&self, left: metapb::Region, right: metapb::Region) -> PdFuture<()>;
+    // For tests.
+    fn get_region_by_id_async(&self, _: u64) -> PdFuture<Option<metapb::Region>> {
+        futures::empty().boxed()
+    }
 }
